refactor: crash handler: move platform specific code to own files
why: clarify sources architecture
