+++ src/platform/platform.h
@@ -2,5 +2,6 @@
 #define PLATFORM_PLATFORM_H
 
 int platform_sdl_version_at_least(int major, int minor, int patch);
+void exit_with_status(int status);
 
 #endif // PLATFORM_PLATFORM_H
